Username String Concatentaion
Type in a certain username and the program will decide what level it belongs to. If the input contains anything other then letter, then it will be level 1. If there are only letters then it will result in level 2. If it also contains more than or equal to 3 vowels, then it will result in level three. Additionally, if the input starts with "a", then it will result in level 4. Finally, if it repeats from the half point like "AvEREaVere" then it will result in level 5.